% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{l62rollcalls}
\alias{l62rollcalls}
\title{Senado LXII Legislatura Info}
\format{An object of class \code{data.frame} with 837 rows and 9 columns.}
\usage{
l62rollcalls
}
\description{
Senado LXII Legislatura Info
}
\keyword{datasets}
