% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{voting_info}
\alias{voting_info}
\title{Voting information}
\usage{
voting_info()
}
\value{
A data.frame with the following information
\itemize{
 \item{\code{ano}}{ Year within term e.g., 'primer ano de ejercicio'}
 \item{\code{tp}}{  Type of session 'ordinario or extraodinario'}
 \item{\code{np}}{  Period within year e.g., 'segundo periodo'}
 \item{\code{lg}}{  Legislature e.g., XVIII}
 \item{\code{id}}{  Compact id including the previous variables}
 \item{\code{description}}{ Original spanish description from web source}
 \item{\code{href}}{ Link to the web source}
}
}
\description{
Creates a data.frame with information about available voting information and web source.
}
\details{
The select names for the output variables are chosen in accordance with the html codes of the web source.
}
